refactor(core): rename interface files to camelCase
