+++ src/core/interfaces/client.ts
@@ -1,4 +1,4 @@
-import type { SendRequestOptions } from './SendRequestOptions';
+import type { SendRequestOptions } from './sendRequestOptions';
 
 export interface Client {
   sendRequest<T>(options: SendRequestOptions<T>): Promise<T>;
