+++ src/core/interfaces/clientConfig.ts
