+++ src/core/interfaces/index.ts
@@ -1,5 +1,3 @@
-export type { Client } from './Client';
-
-export type { ClientConfig } from './ClientConfig';
-
-export type { SendRequestOptions } from './SendRequestOptions';
+export type { Client } from './client';
+export type { ClientConfig } from './clientConfig';
+export type { SendRequestOptions } from './sendRequestOptions';
