+++ src/core/interfaces/sendRequestOptions.ts
